Updated taxonomy, added pdf.
Now there are sections for the types of graph processing we care
about. There are a whole bunch out there so I'm not being complete
yet, but a good format may be something semicolon-delimited, stating
the paper (or a citation eventually), when the project began, its
architecture, last activity, etc.