Added SSSP, PR to the experiments.
Updated el2metis with examples in comments. Removed prepare-compare
because it wasn't doing anything that the .sh files weren't.

Now the experiments have the additional dependencies of a modified
GraphBIG and GraphMat to ensure the tolerances and alpha for PageRank
is consistent. I probably want to make my own forks of those too.