Changed to do 32 runs and added undirected graphs
Now the first NRT runs (I make it 32) are used from the roots file
which is now sorted in increasing order. This is to address the
GraphBIG bug where if you read in N vertices labeled with a
maximum vertex of M with M > N, any root R s.t. N < R < M will
fail.

Changed parsing and dataset generation to account for this.

Added an undirected graph because it might affect GAP
and maybe GraphBIG.