Added graphics to paper and cleaned up tables.
It won't be ready for GABB17 unless I decide I don't need to sleep
for the next 40 hours. Oh well. I tried my best.

Really, what needs to happen next is I need to dust off those old
shell scripts like get-graphalytics, get graphalytics running
again on a wider range of datasets, and then get PowerGraph
working in my own framework and get GraphMat running in
the Graphalytics framework. That way it can be a more interesting
comparison and it can actually show some of the benefits and
limitations of both approaches. Another nice-to-have would be
an improved generation tool for the tables. I think the
autogeneration of the csvs and integration with pgfplotstable
is super cool.

More target architectures would be nice too.

Also, there's still that issue with GraphBIG and its roots in
addition to the fact that some platforms want undirected and some
want directed.